Set up initial CI (#1)
The pipeline performs various code quality and consistency checks such
as:
 - linting with `clippy`
 - checking that the code is formatted according to TBTL's settings
 - checking for unused dependencies, vulnerabilities in used dependencies,
   disallowed licenses, banned crates, etc.
 - building docs
 - building and running tests across various OSs and CPU architectures

Additionally, a script was added to check if all the Rust files have a
header with the proper copyright & license notice.

3rd-party actions were anchored to specific commits so they are used
as immutable releases - mitigating the risk of a bad actor subsequently
adding a backdoor through a branch or tag update.