`bh-jws-utils`: Generalize `Es256SignerWithChain`
This commit generalizes `Es256SignerWithChain` into `SignerWithChain`
which wraps any `Signer` implementation instead of a fixed one.

The existing `Es256SignerWithChain` is turned into a type alias for the
equivalent specialization of the new struct. Certain existing
constructors of it are marked as deprecated due to being superseded by
the new, much more flexible one.

The `openssl` cargo feature is removed, but the gated code is kept,
since:
- the dependency is now necessary for more than just implementing a
  `Signer` backend;
- the gated code had previously silently leaked into a public, non-gated
  function `utils::public_jwk_from_x5chain_leaf`.