`bhx5chain`: Fix `Builder` API oddities
This commit fixes certain issues with the `bhx5chain::Builder` API.

Most prominently, it used to require the leaf _private_ key for
certificate issuing, and then extracted the public key from it. This was
horribly impractical, so the API was relaxed to correctly take the leaf
_public_ key.

Confusing domain-specific wording is improved in documentation and code;
specifically, the word "issuer", which in the Verifiable Credentials
world refers to an end-entity in terms of X.509 certificates, could be
confused for a CA in the X.509 world.

Additionally, types of some arguments are made more idiomatic.

Finally, the purpose of the API is clarified - it is meant to be used as
a lightweight intermediary CA implementation for demos and tests.