`bh-jws-utils`: Bump `bhx5chain` dependency to 0.3.0
This commit bumps the dependency of `bh-jws-utils` to the latest
`bhx5chain` crate.

Previously deprecated `Es256SignerWithChain` constructors which depended
on the old `bhx5chain` API are now removed, as these can be accomplished
with `SignerWithChain` and the new dependency's API.

Additionally, a method to get an `Es256Signer`'s public key in the PEM
format is added. Another method's arguments are refactored to be more
idiomatic.

Some unused 3rd-party dependencies identified by `cargo machete` are
removed.