`bhmdoc`: `ValidityInfo` is no Longer Hardcoded
The `ValidityInfo` information is now passed when issuing credentials,
as opposed to passing the current time and fixing the `signed` and
`validFrom` to the current time, `validUntil` to 1 year after the
current time, and `expectedUpdate` to `null`.

Additionally, the `DateTime` can no longer have sub-second precision,
and requires the use of UTC time zone. `ValidityInfo` can not be
constructed nor deserialized with `validFrom` set to before `signed`,
and `validUntil` set to before `validFrom`.

Fixes: #35